Reorganize project to fit uv needed structure
c++ and python code are moved into their own
respective src folders, since they are largely
independent of each other.
