+++ src/neutralb1/batch/truth_init_ver03.1.cfg
