+++ src/neutralb1/batch/truth_ver03.1.cfg
