+++ src/neutralb1_cpp/batch/extract_bin_info.cc
