+++ src/neutralb1_cpp/batch/extract_cov_matrix.cc
