+++ src/neutralb1_cpp/batch/rand_fit_diagnostic.C
